requirements.txt
# Core dependencies
fastapi==0.115.8
uvicorn==0.25.0
pydantic==2.10.2
python-dotenv==1.0.1

# Database and storage
sqlalchemy==2.0.38
azure-storage-blob==12.16.0
azure-data-tables==12.6.0
azure-core==1.32.0
redis==4.6.0

# Caching and rate limiting
fastapi-cache2[redis]==0.2.2
fastapi-limiter==0.1.6

# Security
PyJWT[crypto]==2.10.1
cryptography>=42.0.0

# HTTP client
httpx==0.25.1

# Monitoring
opencensus>=0.7.0,<1.0.0
opencensus-ext-azure==1.0.1

# Testing
pytest==7.4.3
pytest-mock==3.14.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

azure-identity>=1.10.0
pydantic-settings>=2.0.0
